Use Path::Tiny->tempfile instead of File::Temp
This returns a temporary filename that is unopened. This allows for
Gnuplot to write to the same file since the Perl process will not be
holding on to the filehandle. This is particularly important on Windows
since it only allows a single process to access a file at a time.